Помощь друга. Доработка по замечаниям куратора
